+++ internal/provider/data_source_sdwan_route_policy_definition.go
@@ -234,8 +234,12 @@ func (d *RoutePolicyDefinitionDataSource) Schema(ctx context.Context, req dataso
 										MarkdownDescription: "Metric type",
 										Computed:            true,
 									},
-									"next_hop": schema.StringAttribute{
-										MarkdownDescription: "Next hop IP",
+									"next_hop_prefix_list_id": schema.StringAttribute{
+										MarkdownDescription: "Next hop prefix list ID",
+										Computed:            true,
+									},
+									"next_hop_prefix_list_version": schema.Int64Attribute{
+										MarkdownDescription: "Next hop prefix list version",
 										Computed:            true,
 									},
 									"omp_tag": schema.Int64Attribute{