+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - BREAKING CHANGE: Rename `source_port` attribute to `source_ports` of `sdwan_ipv4_device_acl_policy_definition` resource and data source
 - BREAKING CHANGE: Rename `source_port` attribute to `source_ports` of `sdwan_ipv6_device_acl_policy_definition` resource and data source
 - Add `traffic_class` attribute to `sdwan_ipv6_acl_policy_definition` resource and data source
+- Fix next hop action entry of `sdwan_route_policy_definition` resource and data source
 
 ## 0.3.1